pyproject.toml
[tool.poetry]
name = "biblio"
version = "0.1.0"
description = ""
authors = ["Guillaume TÉCHER <guillaumeval.techer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.44"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
pandas = "^1.5.2"
uvicorn = "^0.20.0"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
